pyproject.toml
[tool.poetry]
name = "thesis"
version = "0.1.0"
description = ""
authors = ["HJacksons <jacksonherberts@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
black = "^24.1.1"
flake8 = "^7.0.0"
wandb = "^0.16.2"
matplotlib = "^3.8.2"
torch = "^2.2.0"
torchvision = "^0.17.0"
scikit-learn = "^1.4.0"
opencv-python = "^4.9.0.80"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
